Refactor gql query typing
