--- .graphqlrc.json
@@ -1,3 +1,0 @@
-{
-  "schema": "http://localhost:4000/graphql"
-}
