+++ next.config.js
@@ -14,14 +14,4 @@ for (const key in process.env) {
   }
 }
 
-nextConfig.webpack = (config, options) => {
-  config.module.rules.push({
-    test: /\.(graphql|gql)/,
-    exclude: /node_modules/,
-    loader: "graphql-tag/loader"
-  });
-
-  return config;
-}
-
 module.exports = nextConfig
